Update operations in CanvasImageData.idl to use long rather than float as per spec https://bugs.webkit.org/show_bug.cgi?id=225263
Reviewed by Simon Fraser.

LayoutTests/imported/w3c:

* canvas/2d.imageData.create2.tiny-expected.txt:
* canvas/2d.imageData.get.tiny-expected.txt:
Update results, which now don't pass, to match the spec behavior.

* web-platform-tests/html/canvas/element/pixel-manipulation/2d.imageData.create2.double-expected.txt:
* web-platform-tests/html/canvas/element/pixel-manipulation/2d.imageData.create2.zero-expected.txt:
* web-platform-tests/html/canvas/element/pixel-manipulation/2d.imageData.get.double-expected.txt:
* web-platform-tests/html/canvas/element/pixel-manipulation/2d.imageData.get.large.crash-expected.txt:
* web-platform-tests/html/canvas/element/pixel-manipulation/2d.imageData.get.rounding-expected.txt:
* web-platform-tests/html/canvas/element/pixel-manipulation/2d.imageData.get.zero-expected.txt:
Update results, which now pass!

Source/WebCore:

Use `[EnforceRange] long` to match the current HTML spec.

* html/canvas/CanvasImageData.idl:
* html/canvas/CanvasRenderingContext2DBase.cpp:
(WebCore::CanvasRenderingContext2DBase::createImageData const):
(WebCore::CanvasRenderingContext2DBase::getImageData const):
(WebCore::CanvasRenderingContext2DBase::putImageData):
* html/canvas/CanvasRenderingContext2DBase.h:

LayoutTests:

* canvas/philip/tests/2d.imageData.create2.tiny-expected.txt:
* canvas/philip/tests/2d.imageData.get.tiny-expected.txt:
Update results, which now don't pass, to match the spec behavior.

* fast/canvas/canvas-2d-imageData-create-nonfinite-expected.txt:
* fast/canvas/canvas-crash.html:
* fast/canvas/canvas-getImageData-invalid-expected.txt:
* fast/canvas/canvas-getImageData-invalid-result-buffer-crash.html:
* fast/canvas/canvas-getImageData-invalid.html:
* fast/canvas/canvas-getImageData-large-crash-expected.txt:
* fast/canvas/canvas-getImageData-rounding-expected.txt:
* fast/canvas/canvas-getImageData-rounding.html:
* fast/canvas/canvas-putImageData-expected.txt:
* fast/canvas/canvas-putImageData.js:
* fast/canvas/resources/canvas-2d-imageData-create-nonfinite.js:
Update tests and results to match spec behavior.


Canonical link: https://commits.webkit.org/237223@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@276877 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/html/canvas/CanvasRenderingContext2DBase.h
@@ -188,10 +188,10 @@ public:
     ExceptionOr<RefPtr<CanvasPattern>> createPattern(CanvasImageSource&&, const String& repetition);
 
     RefPtr<ImageData> createImageData(ImageData&) const;
-    ExceptionOr<RefPtr<ImageData>> createImageData(float width, float height) const;
-    ExceptionOr<RefPtr<ImageData>> getImageData(float sx, float sy, float sw, float sh) const;
-    void putImageData(ImageData&, float dx, float dy);
-    void putImageData(ImageData&, float dx, float dy, float dirtyX, float dirtyY, float dirtyWidth, float dirtyHeight);
+    ExceptionOr<RefPtr<ImageData>> createImageData(int width, int height) const;
+    ExceptionOr<RefPtr<ImageData>> getImageData(int sx, int sy, int sw, int sh) const;
+    void putImageData(ImageData&, int dx, int dy);
+    void putImageData(ImageData&, int dx, int dy, int dirtyX, int dirtyY, int dirtyWidth, int dirtyHeight);
 
     static constexpr float webkitBackingStorePixelRatio() { return 1; }
